♻️ Autoscaling: refactor before 1st draft of computational autoscaling (⚠️ devops) (#4874)
