+++ services/autoscaling/requirements/_test.txt
@@ -106,6 +106,8 @@ httpx==0.24.0
     #   -c requirements/_base.txt
     #   -r requirements/_test.in
     #   respx
+icdiff==2.0.7
+    # via pytest-icdiff
 idna==3.4
     # via
     #   -c requirements/_base.txt
@@ -174,6 +176,8 @@ pbr==5.11.1
     #   sarif-om
 pluggy==1.3.0
     # via pytest
+pprintpp==0.4.0
+    # via pytest-icdiff
 psutil==5.9.5
     # via -r requirements/_test.in
 py-partiql-parser==0.3.6
@@ -200,10 +204,13 @@ pytest==7.4.2
     #   -r requirements/_test.in
     #   pytest-asyncio
     #   pytest-cov
+    #   pytest-icdiff
     #   pytest-mock
 pytest-asyncio==0.21.1
     # via -r requirements/_test.in
 pytest-cov==4.1.0
+    # via -r requirements/_test.in
+pytest-icdiff==0.8
     # via -r requirements/_test.in
 pytest-mock==3.11.1
     # via -r requirements/_test.in
@@ -218,7 +225,9 @@ python-dateutil==2.8.2
 python-dotenv==1.0.0
     # via -r requirements/_test.in
 python-jose==3.3.0
-    # via moto
+    # via
+    #   moto
+    #   python-jose
 pyyaml==6.0.1
     # via
     #   -c requirements/../../../requirements/constraints.txt
