--- services/autoscaling/src/simcore_service_autoscaling/__init__.py
@@ -1,1 +1,0 @@
-from ._meta import __version__
