+++ services/autoscaling/src/simcore_service_autoscaling/_meta.py
@@ -1,35 +1,17 @@
-""" Application's metadata
-
-"""
-from contextlib import suppress
 from typing import Final
 
-import pkg_resources
+from models_library.basic_types import VersionTag
 from packaging.version import Version
+from servicelib.utils_meta import PackageInfo
 
-_current_distribution = pkg_resources.get_distribution("simcore-service-autoscaling")
+info: Final = PackageInfo(package_name="simcore-service-autoscaling")
+__version__: Final[str] = info.__version__
 
-__version__: str = _current_distribution.version
-
-
-APP_NAME: Final[str] = _current_distribution.project_name
-API_VERSION: Final[str] = __version__
-VERSION: Final[Version] = Version(__version__)
-API_VTAG: Final[str] = f"v{VERSION.major}"
-
-
-def get_summary() -> str:
-    with suppress(Exception):
-        try:
-            metadata = _current_distribution.get_metadata_lines("METADATA")
-        except FileNotFoundError:
-            metadata = _current_distribution.get_metadata_lines("PKG-INFO")
-
-        return next(x.split(":") for x in metadata if x.startswith("Summary:"))[-1]
-    return ""  # pragma: no cover
-
-
-SUMMARY: Final[str] = get_summary()
+APP_NAME: Final[str] = info.project_name
+API_VERSION: Final[str] = info.__version__
+VERSION: Final[Version] = info.version
+API_VTAG: Final[VersionTag] = VersionTag(info.api_prefix_path_tag)
+SUMMARY: Final[str] = info.get_summary()
 
 
 # https://patorjk.com/software/taag/#p=testall&f=Avatar&t=Autoscaling
@@ -47,6 +29,24 @@ APP_STARTED_BANNER_MSG = r"""
 )
 
 
-APP_FINISHED_BANNER_MSG = "{:=^100}".format(
-    f"🎉 App {APP_NAME}=={__version__} shutdown completed 🎉"
-)
+APP_STARTED_DYNAMIC_BANNER_MSG = r"""
+      _                                  _
+     | |                                (_)
+   __| | _   _  _ __    __ _  _ __ ___   _   ___
+  / _` || | | || '_ \  / _` || '_ ` _ \ | | / __|
+ | (_| || |_| || | | || (_| || | | | | || || (__
+  \__,_| \__, ||_| |_| \__,_||_| |_| |_||_| \___|
+          __/ |
+         |___/
+"""
+
+APP_STARTED_DISABLED_BANNER_MSG = r"""
+      _  _              _      _            _
+     | |(_)            | |    | |          | |
+   __| | _  ___   __ _ | |__  | |  ___   __| |
+  / _` || |/ __| / _` || '_ \ | | / _ \ / _` |
+ | (_| || |\__ \| (_| || |_) || ||  __/| (_| |
+  \__,_||_||___/ \__,_||_.__/ |_| \___| \__,_|
+"""
+
+APP_FINISHED_BANNER_MSG = info.get_finished_banner()
