+++ services/autoscaling/src/simcore_service_autoscaling/modules/auto_scaling_task.py
@@ -1,17 +1,17 @@
 import json
 import logging
-from typing import Awaitable, Callable
+from collections.abc import Awaitable, Callable
 
 from fastapi import FastAPI
 from servicelib.background_task import start_periodic_task, stop_periodic_task
 from servicelib.redis_utils import exclusive
 
 from ..core.settings import ApplicationSettings
-from .auto_scaling_base import auto_scale_cluster
-from .auto_scaling_dynamic import scale_cluster_with_labelled_services
+from .auto_scaling_core import auto_scale_cluster
+from .auto_scaling_mode_dynamic import DynamicAutoscaling
 from .redis import get_redis_client
 
-_TASK_NAME = "Autoscaling EC2 instances based on docker services"
+_TASK_NAME = "Autoscaling EC2 instances"
 
 logger = logging.getLogger(__name__)
 
@@ -19,13 +19,16 @@ logger = logging.getLogger(__name__)
 def on_app_startup(app: FastAPI) -> Callable[[], Awaitable[None]]:
     async def _startup() -> None:
         app_settings: ApplicationSettings = app.state.settings
-        assert app_settings.AUTOSCALING_NODES_MONITORING  # nosec
-        lock_key = f"{app.title}:cluster_scaling_from_labelled_services_lock"
-        lock_value = json.dumps(
-            {
-                "node_labels": app_settings.AUTOSCALING_NODES_MONITORING.NODES_MONITORING_NODE_LABELS
-            }
-        )
+        lock_key = f"{app.title}:{app.version}:"
+        lock_value = ""
+        if app_settings.AUTOSCALING_NODES_MONITORING:
+            lock_key += f"dynamic:{app_settings.AUTOSCALING_NODES_MONITORING.NODES_MONITORING_NODE_LABELS}"
+            lock_value = json.dumps(
+                {
+                    "node_labels": app_settings.AUTOSCALING_NODES_MONITORING.NODES_MONITORING_NODE_LABELS
+                }
+            )
+        assert lock_value  # nosec
         app.state.autoscaler_task = start_periodic_task(
             exclusive(get_redis_client(app), lock_key=lock_key, lock_value=lock_value)(
                 auto_scale_cluster
@@ -33,7 +36,7 @@ def on_app_startup(app: FastAPI) -> Callable[[], Awaitable[None]]:
             interval=app_settings.AUTOSCALING_POLL_INTERVAL,
             task_name=_TASK_NAME,
             app=app,
-            scale_cluster_cb=scale_cluster_with_labelled_services,
+            auto_scaling_mode=DynamicAutoscaling(),
         )
 
     return _startup
