+++ services/autoscaling/src/simcore_service_autoscaling/modules/ec2.py
@@ -238,7 +238,7 @@ def setup(app: FastAPI) -> None:
             with attempt:
                 connected = await client.ping()
                 if not connected:
-                    raise Ec2NotConnectedError()
+                    raise Ec2NotConnectedError  # pragma: no cover
 
     async def on_shutdown() -> None:
         if app.state.ec2_client:
