+++ services/autoscaling/tests/unit/conftest.py
@@ -3,6 +3,7 @@
 # pylint:disable=redefined-outer-name
 
 import asyncio
+import dataclasses
 import json
 import random
 from collections.abc import AsyncIterator, Awaitable, Callable, Iterator
@@ -34,13 +35,13 @@ from models_library.generated_models.docker_rest_api import (
 )
 from moto.server import ThreadedMotoServer
 from pydantic import ByteSize, PositiveInt, parse_obj_as
-from pytest import MonkeyPatch
 from pytest_mock.plugin import MockerFixture
 from pytest_simcore.helpers.utils_docker import get_localhost_ip
 from pytest_simcore.helpers.utils_envs import EnvVarsDict, setenvs_from_dict
 from settings_library.rabbit import RabbitSettings
 from simcore_service_autoscaling.core.application import create_app
 from simcore_service_autoscaling.core.settings import ApplicationSettings, EC2Settings
+from simcore_service_autoscaling.models import Cluster
 from simcore_service_autoscaling.modules.docker import AutoscalingDocker
 from simcore_service_autoscaling.modules.ec2 import AutoscalingEC2, EC2InstanceData
 from tenacity import retry
@@ -86,7 +87,7 @@ def ec2_instances() -> list[InstanceTypeType]:
 @pytest.fixture
 def app_environment(
     mock_env_devel_environment: EnvVarsDict,
-    monkeypatch: MonkeyPatch,
+    monkeypatch: pytest.MonkeyPatch,
     faker: Faker,
     ec2_instances: list[InstanceTypeType],
 ) -> EnvVarsDict:
@@ -103,13 +104,6 @@ def app_environment(
             "EC2_INSTANCES_SUBNET_ID": faker.pystr(),
             "EC2_INSTANCES_AMI_ID": faker.pystr(),
             "EC2_INSTANCES_ALLOWED_TYPES": json.dumps(ec2_instances),
-            "NODES_MONITORING_NODE_LABELS": json.dumps(["pytest.fake-node-label"]),
-            "NODES_MONITORING_SERVICE_LABELS": json.dumps(
-                ["pytest.fake-service-label"]
-            ),
-            "NODES_MONITORING_NEW_NODES_LABELS": json.dumps(
-                ["pytest.fake-new-node-label"]
-            ),
         },
     )
     return mock_env_devel_environment | envs
@@ -125,6 +119,24 @@ def disable_dynamic_service_background_task(mocker: MockerFixture) -> None:
     mocker.patch(
         "simcore_service_autoscaling.modules.auto_scaling_task.stop_periodic_task",
         autospec=True,
+    )
+
+
+@pytest.fixture
+def enabled_dynamic_mode(
+    app_environment: EnvVarsDict, monkeypatch: pytest.MonkeyPatch
+) -> EnvVarsDict:
+    return app_environment | setenvs_from_dict(
+        monkeypatch,
+        {
+            "NODES_MONITORING_NODE_LABELS": json.dumps(["pytest.fake-node-label"]),
+            "NODES_MONITORING_SERVICE_LABELS": json.dumps(
+                ["pytest.fake-service-label"]
+            ),
+            "NODES_MONITORING_NEW_NODES_LABELS": json.dumps(
+                ["pytest.fake-new-node-label"]
+            ),
+        },
     )
 
 
@@ -426,14 +438,18 @@ def mocked_aws_server() -> Iterator[ThreadedMotoServer]:
     """
     server = ThreadedMotoServer(ip_address=get_localhost_ip(), port=unused_port())
     # pylint: disable=protected-access
-    print(f"--> started mock AWS server on {server._ip_address}:{server._port}")
     print(
-        f"--> Dashboard available on [http://{server._ip_address}:{server._port}/moto-api/]"
+        f"--> started mock AWS server on {server._ip_address}:{server._port}"  # noqa: SLF001
+    )
+    print(
+        f"--> Dashboard available on [http://{server._ip_address}:{server._port}/moto-api/]"  # noqa: SLF001
     )
     server.start()
     yield server
     server.stop()
-    print(f"<-- stopped mock AWS server on {server._ip_address}:{server._port}")
+    print(
+        f"<-- stopped mock AWS server on {server._ip_address}:{server._port}"  # noqa: SLF001
+    )
 
 
 @pytest.fixture
@@ -442,7 +458,7 @@ def reset_aws_server_state(mocked_aws_server: ThreadedMotoServer) -> Iterator[No
     yield
     # pylint: disable=protected-access
     requests.post(
-        f"http://{mocked_aws_server._ip_address}:{mocked_aws_server._port}/moto-api/reset",
+        f"http://{mocked_aws_server._ip_address}:{mocked_aws_server._port}/moto-api/reset",  # noqa: SLF001
         timeout=10,
     )
 
@@ -455,7 +471,7 @@ def mocked_aws_server_envs(
     monkeypatch: pytest.MonkeyPatch,
 ) -> EnvVarsDict:
     changed_envs: EnvVarsDict = {
-        "EC2_ENDPOINT": f"http://{mocked_aws_server._ip_address}:{mocked_aws_server._port}",  # pylint: disable=protected-access
+        "EC2_ENDPOINT": f"http://{mocked_aws_server._ip_address}:{mocked_aws_server._port}",  # pylint: disable=protected-access  # noqa: SLF001
         "EC2_ACCESS_KEY_ID": "xxx",
         "EC2_SECRET_ACCESS_KEY": "xxx",
     }
@@ -576,7 +592,7 @@ async def aws_ami_id(
     ec2_client: EC2Client,
 ) -> str:
     images = await ec2_client.describe_images()
-    image = random.choice(images["Images"])
+    image = random.choice(images["Images"])  # noqa: S311
     ami_id = image["ImageId"]  # type: ignore
     monkeypatch.setenv("EC2_INSTANCES_AMI_ID", ami_id)
     return ami_id
@@ -651,3 +667,21 @@ def fake_ec2_instance_data(faker: Faker) -> Callable[..., EC2InstanceData]:
 async def mocked_redis_server(mocker: MockerFixture) -> None:
     mock_redis = FakeRedis()
     mocker.patch("redis.asyncio.from_url", return_value=mock_redis)
+
+
+@pytest.fixture
+def cluster() -> Callable[..., Cluster]:
+    def _creator(**cluter_overrides) -> Cluster:
+        return dataclasses.replace(
+            Cluster(
+                active_nodes=[],
+                drained_nodes=[],
+                reserve_drained_nodes=[],
+                pending_ec2s=[],
+                disconnected_nodes=[],
+                terminated_instances=[],
+            ),
+            **cluter_overrides,
+        )
+
+    return _creator
