+++ services/director-v2/tests/unit/with_dbs/test_modules_comp_scheduler_dask_scheduler.py
@@ -1056,6 +1056,7 @@ class RebootState:
     expected_run_state: RunningState
 
 
+@pytest.mark.flaky(max_runs=3)
 @pytest.mark.parametrize(
     "reboot_state",
     [
